Refactor & remove user dependency
